feat: separate Mermaid ERD into dedicated file and extract MermaidService (v2.9.0)
- Extract Mermaid ERD logic from SOQLTemplateService into new MermaidService
- DirectoryProcessor now writes a dedicated mermaid-erd--{tree}-{timestamp}.md
  file alongside the soql-sosl-templates file for each recipe tree
- Mermaid file includes title, description of diagram notation, timestamp, and object list
- SOQL/SOSL template file no longer contains the Entity Relationship Diagram section
- Add Required Field Identification and Documentation task to backlog
- Bump version to 2.9.0
