+++ package.json
@@ -3,7 +3,7 @@
   "displayName": "Salesforce Data Treecipe",
   "description": "source-fidelity driven development, pairs well with cumulus-ci",
   "icon": "images/datatreecipe.webp",
-  "version": "2.8.0",
+  "version": "2.9.0",
   "engines": {
     "vscode": "^1.94.0"
   },
