+++ src/treecipe/src/SOQLTemplateService/SOQLTemplateService.ts
@@ -1,7 +1,6 @@
 import { ObjectInfoWrapper } from '../ObjectInfoWrapper/ObjectInfoWrapper';
 import { ObjectInfo } from '../ObjectInfoWrapper/ObjectInfo';
 import { FieldInfo } from '../ObjectInfoWrapper/FieldInfo';
-
 const LOOKUP_FIELD_TYPES = ['Lookup', 'MasterDetail', 'Hiearchy'];
 
 const TEXT_FIELD_TYPES = ['Text', 'TextArea', 'LongTextArea', 'Html', 'Email', 'Phone', 'Url'];
@@ -32,14 +31,6 @@ export class SOQLTemplateService {
             `Objects: ${objectNames.join(', ')}`,
             ``,
             `---`,
-            ``,
-            `## Entity Relationship Diagram`,
-            ``,
-            `\`\`\`mermaid`,
-            this.buildMermaidERD(objectInfoWrapper),
-            `\`\`\``,
-            ``,
-            `---`,
         ];
 
         for (const objectName of objectNames) {
@@ -49,52 +40,6 @@ export class SOQLTemplateService {
         }
 
         return sections.join('\n');
-
-    }
-
-    static buildMermaidERD(objectInfoWrapper: ObjectInfoWrapper): string {
-
-        const lines: string[] = ['erDiagram'];
-        const objectNames = Object.keys(objectInfoWrapper.ObjectToObjectInfoMap).sort();
-
-        for (const objectName of objectNames) {
-            const objectInfo = objectInfoWrapper.ObjectToObjectInfoMap[objectName];
-            const sanitizedName = this.sanitizeForMermaid(objectName);
-
-            lines.push(`    ${sanitizedName} {`);
-            lines.push(`        id Id`);
-
-            const nonRelationshipFields = (objectInfo.Fields ?? []).filter(
-                f => !LOOKUP_FIELD_TYPES.includes(f.type)
-            );
-            for (const field of nonRelationshipFields) {
-                const mermaidType = this.getMermaidFieldType(field.type);
-                lines.push(`        ${mermaidType} ${field.fieldName}`);
-            }
-
-            lines.push(`    }`);
-        }
-
-        lines.push(``);
-
-        for (const objectName of objectNames) {
-            const objectInfo = objectInfoWrapper.ObjectToObjectInfoMap[objectName];
-            for (const field of (objectInfo.Fields ?? [])) {
-
-                if (!LOOKUP_FIELD_TYPES.includes(field.type)) { continue; }
-
-                const parentName = field.referenceTo;
-                if (!parentName || !(parentName in objectInfoWrapper.ObjectToObjectInfoMap)) { continue; }
-
-                const parentSanitized = this.sanitizeForMermaid(parentName);
-                const childSanitized = this.sanitizeForMermaid(objectName);
-                const cardinality = field.type === 'MasterDetail' ? `||--o{` : `|o--o{`;
-                lines.push(`    ${parentSanitized} ${cardinality} ${childSanitized} : "${field.fieldName}"`);
-
-            }
-        }
-
-        return lines.join('\n');
 
     }
 
@@ -291,39 +236,6 @@ export class SOQLTemplateService {
         }
 
         return `${childObjectName}s`;
-
-    }
-
-    static sanitizeForMermaid(objectName: string): string {
-        return objectName.replace(/-/g, '_');
-    }
-
-    static getMermaidFieldType(salesforceFieldType: string): string {
-
-        const typeMap: Record<string, string> = {
-            'Text': 'string',
-            'TextArea': 'string',
-            'LongTextArea': 'string',
-            'Html': 'string',
-            'Email': 'string',
-            'Phone': 'string',
-            'Url': 'string',
-            'Number': 'number',
-            'Currency': 'number',
-            'Percent': 'number',
-            'Date': 'date',
-            'DateTime': 'datetime',
-            'Boolean': 'boolean',
-            'Checkbox': 'boolean',
-            'Picklist': 'string',
-            'MultiselectPicklist': 'string',
-            'MultiSelectPicklist': 'string',
-            'AutoNumber': 'string',
-            'Formula': 'string',
-            'EncryptedText': 'string',
-        };
-
-        return typeMap[salesforceFieldType] ?? 'string';
 
     }
